backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
loguru==0.7.2
google-cloud-firestore==2.16.0

# AI Provider SDKs
google-generativeai==0.8.3
openai==1.54.5
groq==0.11.0
aiohttp==3.10.11
